Import proprietary code under specialized license
