+++ app/src/main/java/com/hiddenramblings/tagmo/Util.java
@@ -30,10 +30,6 @@ public class Util {
 
     public static final String AMIIBO_DATABASE_FILE = "amiibo.json";
 
-    private static final String STORAGE_ROOT = "/storage";
-
-    private static File storagePath;
-
     public static String bytesToHex(byte[] bytes) {
         StringBuilder sb = new StringBuilder();
         for (byte b : bytes) {
@@ -121,52 +117,8 @@ public class Util {
         }
     }
 
-    @SuppressWarnings("ConstantConditions")
-    private static File getRootPath(File directory) {
-        return directory.getParentFile().getParentFile().getParentFile().getParentFile();
-    }
-
-    @RequiresApi(api = Build.VERSION_CODES.LOLLIPOP)
-    private static File setFileStorageLollipop() {
-        File[] storage = ContextCompat.getExternalFilesDirs(TagMo.getContext(), null);
-        if (TagMo.getPrefs().ignoreSdcard().get())
-            return storagePath = storage[0] != null && storage[0].canRead()
-                    ? getRootPath(storage[0]) : Environment.getExternalStorageDirectory();
-        try {
-            File sdcard = storage.length > 1 && storage[1] != null
-                    && !isExternalStorageEmulated(storage[1]) ? storage[1] : storage[0];
-            // [TARGET]/Android/data/[PACKAGE]/files
-            return storagePath = sdcard.canRead() ? getRootPath(sdcard)
-                    : Environment.getExternalStorageDirectory();
-        } catch (IllegalArgumentException | NullPointerException e) {
-            return storagePath = Environment.getExternalStorageDirectory();
-        }
-    }
-
-    @SuppressWarnings("ConstantConditions")
-    public static File setFileStorage() {
-        File emulated = null;
-        File physical = null;
-        try {
-            for (File directory : new File(STORAGE_ROOT).listFiles()) {
-                if (directory.getAbsolutePath().endsWith("emulated"))
-                    emulated = directory;
-                else if (!directory.getAbsolutePath().endsWith("self"))
-                    physical = directory;
-            }
-        } catch (NullPointerException e) {
-            if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP)
-                setFileStorageLollipop();
-            else
-                return storagePath = Environment.getExternalStorageDirectory();
-        }
-        if (TagMo.getPrefs().ignoreSdcard().get())
-            return storagePath = emulated != null ? emulated : physical;
-        return storagePath = physical != null ? physical : emulated;
-    }
-
     public static File getSDCardDir() {
-        return storagePath != null ? storagePath : setFileStorage();
+        return Storage.getStorageFile();
     }
 
     public static File getFilesDir() {
